@@ -1,68 +1,56 @@
 {
   "name": "ci",
-  "version": "0.9.0",
+  "version": "1.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve  --open --base-href /gizmo/",
     "build": "ng build --output-path=last-build/ --env=prod --base-href /gizmo/",
-    "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
     "prod-build": "ng build --output-path=last-build/ --env=prod --base-href /gizmo/"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "^5.2.10",
-    "@angular/compiler": "^5.2.10",
-    "@angular/core": "^5.2.10",
-    "@angular/flex-layout": "5.0.0-beta.14",
-    "@angular/forms": "^5.2.10",
-    "@angular/http": "^5.2.10",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
+    "@angular/animations": "^6.0.5",
+    "@angular/cdk": "^6.2.1",
+    "@angular/common": "^6.0.5",
+    "@angular/compiler": "^6.0.5",
+    "@angular/core": "^6.0.5",
+    "@angular/flex-layout": "^6.0.0-beta.16",
+    "@angular/forms": "^6.0.5",
+    "@angular/http": "^6.0.5",
+    "@angular/material": "^6.2.1",
+    "@angular/platform-browser": "^6.0.5",
+    "@angular/platform-browser-dynamic": "^6.0.5",
+    "@angular/router": "^6.0.5",
     "@types/jquery": "^3.3.1",
-    "@types/mathjax": "0.0.32",
+    "@types/mathjax": "0.0.34",
     "angular-sortablejs": "^2.5.2",
     "angular2-fontawesome": "^5.2.1",
-    "copy-webpack-plugin": "4.5.0",
-    "core-js": "^2.5.5",
+    "copy-webpack-plugin": "4.5.1",
+    "core-js": "^2.5.7",
     "expr-eval": "^1.2.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
-    "mathjax": "^2.7.2",
-    "rxjs": "^5.5.0",
+    "mathjax": "2.7.2",
+    "rxjs": "^6.2.0",
     "sortablejs": "^1.7.0",
-    "systemjs": "^0.21.3",
+    "systemjs": "^0.21.4",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
-    "@types/node": "~9.4.6",
+    "@angular/cli": "^6.0.5",
+    "@angular/compiler-cli": "^6.0.5",
+    "@angular/language-service": "^6.0.5",
+    "@types/node": "~10.3.3",
     "codelyzer": "~4.3.0",
-    "ts-node": "~6.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "2.4.2",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2"
+    "ts-node": "~6.1.1",
+    "tslint": "~5.10.0",
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.8"
   },
   "optionalDependencies": {
-    "fsevents": "^1.2.2"
+    "fsevents": "^1.2.4"
   }
 }
