@@ -1,6 +1,7 @@
 ﻿import { Component } from '@angular/core';
 import { Router, NavigationEnd, ActivatedRoute } from '@angular/router';
-import 'rxjs/add/operator/filter';
+import { map, filter } from 'rxjs/operators';
+import { HTTPStatus } from '../_services/index';
 
 @Component({
     moduleId: module.id,
@@ -10,6 +11,8 @@ import 'rxjs/add/operator/filter';
 })
 
 export class AppComponent {
+    HTTPActivity: boolean;
+
     public showMenu = this.isLoggedIn();
 
     protected isLoggedIn() {
@@ -18,17 +21,22 @@ export class AppComponent {
 
     constructor(
         private router: Router,
-        private activatedRoute: ActivatedRoute) {
+        private activatedRoute: ActivatedRoute,
+        private httpStatus: HTTPStatus) {
+            this.httpStatus.getHttpStatus()
+                .subscribe((status: boolean) => {this.HTTPActivity = status;});
             router.events
-                .filter((event) => event instanceof NavigationEnd)
-                .map(() => activatedRoute)
-            .subscribe((event) => {
-                if (router.url == "/login") {
-                    this.showMenu = false;
-                }
-                else {
-                    this.showMenu = this.isLoggedIn();
-            }
-            });
+                .pipe(
+                    filter((event) => event instanceof NavigationEnd),
+                    map(() => activatedRoute)
+                )
+                .subscribe((event) => {
+                    if (router.url == "/login") {
+                        this.showMenu = false;
+                    }
+                    else {
+                        this.showMenu = this.isLoggedIn();
+                    }
+                });
     };
 }
