@@ -1,7 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions, Response } from '@angular/http';
 import { Observable } from 'rxjs';
-import 'rxjs/add/operator/map';
 
 import { HttpService } from './http.service';
 
@@ -11,25 +9,21 @@ export class PlacementService {
 	constructor(private http: HttpService) { }
 
 	getPlacementQuestions() {
-	    return this.http.get('/placement')
-	        .map((response: Response) => response);
+	    return this.http.get('/placement');
 	}
 
 	getFirstTopicId(unitId) {
-		return this.http.get('/placement/getTopicId/'+unitId)
-	        .map((response: Response) => response);
+		return this.http.get('/placement/getTopicId/'+unitId);
 	}
 
 	doneUnit(unitId) {
-		let request = JSON.stringify({ unit_id: unitId });
-		return this.http.post('/placement/done-unit', request)
-	        .map((response: Response) => response);
+		let request = { unit_id: unitId };
+		return this.http.post('/placement/done-unit', request);
 	}
 
 	doneHalfUnit(unitId) {
-		let request = JSON.stringify({ unit_id: unitId });
-		return this.http.post('/placement/done-half-unit', request)
-	        .map((response: Response) => response);
+		let request = { unit_id: unitId };
+		return this.http.post('/placement/done-half-unit', request);
 	}
 
 }
